Refactor test structure
