+++ templates/Virtual-Assistant-Template/typescript/generator-botbuilder-assistant/generators/skill/src/copier.js
@@ -79,8 +79,8 @@ class Copier {
       path.join(`src`, `dialogs`, `skillDialogBase.ts`)
     );
     templateFiles.set(
-      path.join(`test`, `mockResources`, `_cognitiveModels.json`),
-      path.join(`test`, `mockResources`, `cognitiveModels.json`)
+      path.join(`test`, `mocks`, `resources`, `_cognitiveModels.json`),
+      path.join(`test`, `mocks`, `resources`, `cognitiveModels.json`)
     );
     selectedLanguages.forEach(language => {
       templateFiles.set(
