+++ templates/Virtual-Assistant-Template/typescript/generator-botbuilder-assistant/generators/skill/templates/customSkill/_package.json
@@ -52,8 +52,6 @@
         "nyc": "^14.1.1",
         "replace": "^1.0.0",
         "rimraf": "^2.6.2",
-        "source-map-support": "^0.5.12",
-        "ts-node": "^8.1.0",
         "tslint": "^5.12.1",
         "tslint-microsoft-contrib": "6.0.0",
         "typescript": "^3.2.2"
