+++ templates/Virtual-Assistant-Template/typescript/generator-botbuilder-assistant/generators/skill/templates/customSkill/test/helpers/testBase.js
@@ -8,7 +8,7 @@ const join = require("path").join;
 
 const nockBack = require("nock").back;
 nockBack.setMode(TEST_MODE);
-nockBack.fixtures = join(__dirname, "/nockFixtures");
+nockBack.fixtures = join(__dirname, '..', 'mocks', 'nockFixtures');
 
 const uuidRegex = /[a-f\d-]{8}-[a-f\d-]{4}-[a-f\d-]{4}-[a-f\d-]{4}-[a-f\d-]{12}/;
 const qnaRegex = /\/\/[^.]+\.azurewebsites/;
