+++ templates/Virtual-Assistant-Template/typescript/generator-botbuilder-assistant/generators/skill/templates/customSkill/test/interruption.test.js
@@ -3,8 +3,8 @@
  * Licensed under the MIT License.
  */
 
-const skillTestBase = require("./skillTestBase");
-const testNock = require("../testBase");
+const skillTestBase = require("./helpers/skillTestBase");
+const testNock = require("./helpers/testBase");
 
 describe("interruption", function() {
   beforeEach(async function() {
