+++ templates/Virtual-Assistant-Template/typescript/generator-botbuilder-assistant/generators/skill/templates/customSkill/test/localization.test.js
@@ -3,8 +3,8 @@
  * Licensed under the MIT License.
  */
 
-const skillTestBase = require("./skillTestBase");
-const testNock = require("../testBase");
+const skillTestBase = require("./helpers/skillTestBase");
+const testNock = require("./helpers/testBase");
 
 describe("localization", function() {
   beforeEach(async function() {
