+++ templates/Virtual-Assistant-Template/typescript/generator-botbuilder-assistant/generators/skill/templates/customSkill/test/mocha.opts
@@ -1,7 +1,4 @@
---require ts-node/register
---require source-map-support/register
 --reporter mocha-junit-reporter
 --timeout 50000
 --recursive
 --colors
-**/*Test.js
