--- templates/Virtual-Assistant-Template/typescript/generator-botbuilder-assistant/generators/skill/templates/customSkill/test/mockResources/appsettings.json
@@ -1,25 +1,0 @@
-{
-    "oauthConnections": [
-      {
-        "name": "",
-        "provider": ""
-      }
-    ],
-    "properties": {},
-    "microsoftAppId": "00000002-0000-0000-0000-000000000001",
-    "microsoftAppPassword": "00000002-0000-0000-0000-000000000001",
-    "appInsights": {
-      "appId": "skill-appInsights-applicationId",
-      "instrumentationKey": "skill-appInsights-instrumentationKey"
-    },
-    "blobStorage": {
-      "connectionString": "skill-blob-connectionString",
-      "container": "transcripts"
-    },
-    "cosmosDb": {
-      "authkey": "skill-cosmosdb-key",
-      "collectionId": "botstate-collection",
-      "cosmosDBEndpoint": "https://skill-endpoint.documents.azure.com:443/",
-      "databaseId": "botstate-db"
-    }
-  }
