+++ templates/Virtual-Assistant-Template/typescript/generator-botbuilder-assistant/generators/skill/templates/customSkill/test/mocks/nockFixtures/localization_de_response.json
