+++ templates/Virtual-Assistant-Template/typescript/generator-botbuilder-assistant/generators/skill/templates/customSkill/test/mocks/resources/appsettings.json
@@ -1,0 +1,25 @@
+{
+  "oauthConnections": [
+    {
+      "name": "",
+      "provider": ""
+    }
+  ],
+  "properties": {},
+  "microsoftAppId": "<--MOCK-VALUE-->",
+  "microsoftAppPassword": "<--MOCK-VALUE-->",
+  "appInsights": {
+    "appId": "<--MOCK-VALUE-->",
+    "instrumentationKey": "<--MOCK-VALUE-->"
+  },
+  "blobStorage": {
+    "connectionString": "<--MOCK-VALUE-->",
+    "container": "transcripts"
+  },
+  "cosmosDb": {
+    "authkey": "<--MOCK-VALUE-->",
+    "collectionId": "botstate-collection",
+    "cosmosDBEndpoint": "https://skill-endpoint.documents.azure.com:443/",
+    "databaseId": "botstate-db"
+  }
+}
